Feat: swap Exact Amount Out, SQS Routing and API changes
Create a common Quote class for reusing methods across different swap
methods tests.
